Separate ryotei data formatting logic from useGetRyotei hook
