+++ feature/ryotei/hooks/useGetRyotei.ts
@@ -1,11 +1,8 @@
 'use client'
-import { Plan } from '@/component'
 import { QUERY_GET_RYOTEI } from '@/feature/ryotei/graphql'
 import { GetRyoteiQuery, GetRyoteiQueryVariables, OrderByDirection } from '@/feature/api/graphql'
 import { useQuery } from '@apollo/client/react'
-import { format } from 'date-fns'
-import { TZDate } from '@date-fns/tz'
-import { ja } from 'date-fns/locale/ja'
+import { formatRyoteiData } from '@/feature/ryotei/utils/formatRyoteiData'
 
 export const useGetRyotei = (selectedTripId?: string) => {
   const variables: GetRyoteiQueryVariables = {
@@ -16,28 +13,11 @@ export const useGetRyotei = (selectedTripId?: string) => {
     variables,
   })
 
-  const res = data?.ryoteiCollection?.edges?.map((edge) => edge.node)
-  const grouped: Record<string, Plan[]> | undefined = res?.reduce((acc: Record<string, Plan[]>, node) => {
-    const { id, datetime, description, trip_id } = node
-    const rowTZDate = new TZDate(datetime + 'Z')
-    const formatDate = format(rowTZDate, 'yyyy-MM-dd', { locale: ja })
-    const formatRowDate = format(rowTZDate, "yyyy-MM-dd'T'HH:mm:ss.SS", { locale: ja })
-    if (!acc[formatDate]) {
-      acc[formatDate] = []
-    }
-    acc[formatDate].push({ datetime: formatRowDate, description, id, trip_id })
-    acc[formatDate].sort((a, b) => a.datetime.localeCompare(b.datetime))
-    return acc
-  }, {})
+  const nodes = data?.ryoteiCollection?.edges?.map((edge) => edge.node)
+  const formattedData = formatRyoteiData(nodes)
+
   return {
-    data: grouped
-      ? Object.keys(grouped)
-          .sort()
-          .reduce((acc: Record<string, Plan[]>, key: string) => {
-            acc[key] = grouped[key]
-            return acc
-          }, {})
-      : undefined,
+    data: formattedData,
     refetch,
   }
 }
